added more base/venue css, js
