+++ gighubclient/src/App.css
@@ -2,29 +2,33 @@
   text-align: center;
 }
 
-.App-logo {
-  height: 20vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
 
 .App-header {
   background-color: #22194D;
   min-height: 33vh;
   display: flex;
-  flex-direction: column;
+  flex-direction: row;
   align-items: center;
-  justify-content: center;
+  justify-content: left;
   font-size: calc(10px + 2vmin);
   color: white;
+  padding: 15px 15px 15px 80px;
+  
 }
 
-.App-link {
+.App-logo {
+  height: 20vmin;
+  pointer-events: none;
+}
+
+/* @media (prefers-reduced-motion: no-preference) {
+  .App-logo { */
+    /* animation: App-logo-spin infinite 20s linear; */
+    /* margin: 20px; */
+  /* }
+} */
+
+/* .App-link {
   color: #61dafb;
 }
 
@@ -35,4 +39,13 @@
   to {
     transform: rotate(360deg);
   }
+} */
+
+.App-title {
+  margin: 10px;
+  font-family: 'Jockey One';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 115px;
+  line-height: 161px;
 }
