+++ gighubclient/src/components/venue/VenueList.css
@@ -1,0 +1,34 @@
+.venue-list {
+    background: #D9D9D9;
+    border: black solid;
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    justify-content: space-evenly;
+    margin: 20px;
+}
+
+.page-title {
+    font-family: 'Jockey One';
+    font-style: normal;
+    font-size: 60px;
+    text-align: center;
+    margin: 20px;
+}
+
+.venue-card {
+    background: #F5F5F5;
+    border: 5px solid black;
+    border-radius: 5px;
+    margin: 20px;
+    padding: 0 10px;
+    max-width: min-content;
+}
+
+.venue-details {
+    list-style-type: none;
+}
+
+.venue-user-details {
+    list-style-type: none;
+}
