+++ gighubclient/src/index.css
@@ -1,7 +1,5 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
 body {
-  background-color: rgb(245, 245, 245)
+  background-color: rgb(245, 245, 245);
+  font-family: 'Jockey One', sans-serif;
+  font-family: 'Merriweather', serif;
 }
